make the structure of files more clear, reuse the exception code
