+++ protostar/cairo/bindings/cairo_bindings.py
@@ -7,12 +7,7 @@ from contextlib import contextmanager
 import cairo_python_bindings
 from protostar.cairo.cairo_function_runner_facade import RUNNER_BUILTINS_TITLE_CASE
 
-
-class CairoBindingException(Exception):
-    def __init__(self, message: str, details: Optional[str] = None):
-        self.message = message
-        self.details = details
-        super().__init__(message)
+from .cairo_bindings_exception import CairoBindingException
 
 
 def ensure_output_path(output_path: Optional[Path]):
