+++ protostar/cairo/bindings/cairo_bindings_exception.py
@@ -1,0 +1,8 @@
+from typing import Optional
+
+
+class CairoBindingException(Exception):
+    def __init__(self, message: str, details: Optional[str] = None):
+        self.message = message
+        self.details = details
+        super().__init__(message)
