+++ tests/integration/cairo1/misc/cairo_runner_facade_test.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-import protostar.cairo.cairo_bindings as cairo1
+import protostar.cairo.bindings.cairo_bindings as cairo1
 from protostar.cairo.cairo1_test_suite_parser import (
     ProtostarCasm,
 )
