+++ tests/integration/cairo1/rust_test_runner_bindings/rust_test_runner_bindings_test.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from protostar.cairo import protostar_rust_bindings
+from protostar.cairo.bindings import protostar_rust_bindings
 
 
 async def test_print(datadir: Path):
